修正 每日計算訓練記錄、每週訓練記錄計算邏輯

+++ PhysicalFit/Models/AthleteArcheryTrainingRecord.cs
@@ -23,7 +23,7 @@ namespace PhysicalFit.Models
         public Nullable<int> Poundage { get; set; }
         public Nullable<int> ArrowCount { get; set; }
         public Nullable<int> RPEscore { get; set; }
-        public Nullable<int> EachTrainingLoad { get; set; }
+        public Nullable<decimal> EachTrainingLoad { get; set; }
         public Nullable<int> DailyTrainingLoad { get; set; }
         public Nullable<int> SessionRPEAthleteRecordID { get; set; }
